feat: camera mvmt, first impl grid, simple agent
Camera Movement: you can go up and down, left right and zoom in with scroll wheel.
First grid: amount of cells in width and height are adjustable before generating. Each cell has a random direction. Every odd cell is gray for visual clarity.
Simple Agent: accesses the grid to see which direction the cell he is on is, moves that direction